sub.in/install2/Makefile: enforce a kernel
A bit earlier the situation where there was a stage1 fallback
for INSTALLER_KFLAVOUR as the last of KFLAVOURS, so we'd have
an installer kernel or build would break; now the situration
got somewhat twisted: there could be a vmlinuz in stage1 but
no corresponding modules in install2 -- which can lead to
different surprises but at least alterator-vm would complain
about "Operation not permitted" on partition layout commit.

The fallback is a copypaste from sub.in/stage1/Makefile though
and should be redone properly somewhere.  The question so far
is, where exactly?