regular.mk: test refind with regular-icewm.iso
The current refind support implementation doesn't boot
with either USB Flash or CD-ROM for me (due to the different
actual volumes, FAT vs El Torito, accordingly).

And when it does, a host with Secure Boot present and not
disabled yet would turn the unsigned kernel down unlike ELILO
(due to rEFInd using "natural" calls to boot it).

Both issues are reported upstream; in the meanwhile let's
migrate a single test image but build the rest with elilo.