net: avoid NM showstopper at boot by default
Intro: NetworkManager-wait-online.service would, well, wait
for some network interface to become online or for timeout
to kick in.

Problem: if a LiveCD is tested in offline environment
that timeout will only impede the boot.

Proposed solution: use/net/nm/nodelay target has been implemented
to disable that service as proposed by sem@ and done in Simply;
"+nm" target changed to be an alias to this one.