armh: get rid of 50-sshd-root
It was a temporary hack actually, and is better dropped long-term:
things like predefined root accounts with remote access are *evil*
and this hook was a half of that "solution".

Use of oem feature to integrate first-boot setup is recommended
to deal with this issue, at least when graphics are available.