rescue+*: added assorted extra packages
It was the proposal to add fsarchiver that has started this;
the package was there in X11-bearing live-rescue.iso but appears
to be a console program thus moved to rescue+misc pkglist;
more than a few rescue-grade utilities have turned up during
a quick look at what else is missing.

shellinabox and dvdsaster have been "added" as candidates since
the former does require additional actions but can provide a nice
security hole if these are taken without extra consideration,
and the latter is just pretty large although might still be useful.