desktop+gnustep: app status update
Split development packages into dev+gnustep pkglist -- these are
worth including in "full" version but will need thorough testing
so as to present the tools to those who value these.

Some of user packages are problematic and shouldn't be included
right now; the problems are mostly of these kinds:

- app won't start (at all or effectively);
- useless for being too alpha quality/incomplete;
- menu file for a commandline app lacking any feedback;
- package lacks the dependencies needed;
- it's a LoginPanel ;-)

Thanks a lot to upstream authors, real@ the packager
and kostyalamer who prepared a lot of menufiles anyways!