p/l/t/README: updates and tweaks
Some more editing has been due over pkg.in/lists/tagged/README
to make it more comprehensible and up-to-date; the problem with
groups isn't actually that bad as alterator-pkg's groups concept
is currently aligned with the requisite functionality provided by
pkg.in/lists/* directly; the tagged pkglists come into play when
we want to add "something like that" and don't really care about
the fine details of a secondary thing trusting that it's actually
comprised and working as advertized through its name tags.

Compare to reusing the pre-existing image configuration or features
versus reimplementing things in a rigid manner -- it's a flexibility
vs predictability question, and both scenarios are supported within
m-p explicitly.