conf.d: updated README to mention mixin/*
The mixin concept and name has been borrowed from Ruby
language -- it's a kind of thing that can be added to
more or less whatever suitable; the problem it tries
to solve is that incrementally building up the image
configuration breaks when one would like to change
something that's been configured in early enough so that
grafting early will warrant a lot of duplication later on
but inheriting too much things that need to be changed
gets too much hackery in.

It started while trying to build an installer image
using configuration bits and pieces collected while
bringing an installable LiveCD together: there are
just too many livecdish things in a LiveCD to try
and rebase the actual desktop configuration things
onto an installer, so putting these into a mixin
to be reused within both livecd and installer
seems the way to go.