initial browser feature
This one has been asking to be implemented for too long already,
and zerg@ was interested in a bit more lean and mean regular-kde4
either (there are two browsers provided with it via metapackage).

There's another reason to do it recently: Firefox Australis UI
is not exactly the best for many of us, and good ol' seamonkey
seems preferable for "vintage"/low-resource images coming with
icewm or windowmaker.