init: added tty12 log setup for systemd case
That's a part of ALT Linux conveniences: system log
messages at tty12 (helps immensely in case of disk crash
or cable problems as running anything, including utilities
to view logs, becomes painful to impossible in such cases).

systemd lacks this kind of setup out-of-box for sure
so zerg@ hacked a substitute together; just pull that in.