fonts: add a few font collection targets
Font packages are sprinkled all over the metaprofiles,
let's try and help make their use more systematic.

This is a sort of a feature abuse as it was conceived
for fontconfig setup originally but spawning features
with confusing names looks grim; so let all things fonts
live within a feature named "fonts" for the time being.