exterminate fonts-ttf-dejavu all over the place
It looks *ugly* on-screen, at least within regular builds,
even if the screen is 166dpi.

Based on a quick experiment this morning I'd suggest using
fonts-otf-adobe-source-{code,sans}-pro instead -- and it's
available as use/fonts/otf/adobe now, incidentally.

The documentation is still built with it though as a2x/fop look
unhappy otherwise (as in replacing Cyrillic glyphs with "#"s).