init: ban systemd when sysvinit has been requested
The issue at hand is that interactivesystem pulls in
network-config-subsystem and that one has several providers
by now, systemd-networkd being one of them since recently
(and pulling in systemd).

Just the same problem as with systemd-journal; both might be
fixed by reworking mkimage to allow for different package name
resolution modes:

- "slap everything together and resolve with one-shot"
  to handle conflictless situations (most of those);
- "process multiple transactions to allow for conflicts"
  thus making it possible to include e.g. a few MTAs into
  the provided package base.

Ensure that systemd is outside by explicitly telling so
in the pkglist.