regular.mk: added xpra and winswitch to icewm
xpra.org and winswitch.org are wonderful applications
helping one working with X11 apps all over the place,
running those on local and remote desktops and servers
with persistent sessions.

These aren't exactly trivial to plan and use but icewm
isn't for newbies either; thus these seem like nice match
for an "X11 thin client" image.

Size addition is ca. 50 Mb as of today which is a lot
but worth it.