regular.mk, regular+tde: be careful with kpowersave
I don't know of [k]powersave interactions with systemd
and the rest of that ilk, let's be cautious and only provide
it for sysv-tde images which have been actually tested
by Speccyfighter who asked for this package.

Can't tell for sure but something seems to break
suspend menu option (hibernate isn't supported in
all of the regular builds/starterkits just yet).