syslinux: added 84rescue_remote.cfg
This one provides cmdline arguments for startup-rescue >= 0.24
which would bring up networking and sshd in its turn thus allowing
remote access to the host booted in this mode.

The feature has been asked for by many people including mithraen@
and valintinr@ (and I'd make use of it another day too).

See the appropriate startup-rescue commit description for notes
on implementation; this default set of variable values should be
both useful and illustrative though.