stage2: mv use/stage2/cifs{-install,}
This has no users in master but out-of-tree branches might need
a trivial update.

The rationale is that it's actually for *any* stage2 and not related
to specifically "install" at all (otherwise it should have been moved
to install2 feature altogether).

Note that there's no reason to add nfs-utils similarly as make-initrd
requires kinit-utils which includes its own nfsmount.